fix: fix issues with authjs library

+++ apps/webapp/src/routes/org/[organizationId]/settings/members/[memberId]/+page.server.ts
@@ -1,15 +1,17 @@
 import { getSelectedOrganization } from '$lib/server/action-utils';
 import { OrganizationService } from '$lib/server/organization';
-import type { Session } from '@auth/core/types';
+import { error, redirect } from '@sveltejs/kit';
 import type { Actions } from './$types';
-import { redirect } from '@sveltejs/kit';
 
 export const actions: Actions = {
 	/**
 	 * Remove a member from the organization.
 	 */
 	revokeMembership: async (event) => {
-		const session = (await event.locals.auth()) as Session;
+		const session = await event.locals.auth();
+		if (!session || !session.user?.id) {
+			return error(403);
+		}
 		const organization = await getSelectedOrganization(event);
 		const userId = event.params.memberId;
 		await OrganizationService.removeMember(organization.id, userId);
